requirements.txt
blinker==1.7.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gpio==1.0.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
oauthlib==3.2.2
packaging==25.0
pdfkit==1.0.0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyMySQL==1.1.0
pyodbc==5.3.0
pyparsing==3.2.1
python-dotenv==1.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.27
sqlalchemy-iris==0.17.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.0.1
WTForms==3.1.2
pytest==7.4.0
pytest-cov==4.1.0
unittest-xml-reporting==3.2.0